feat: authentication for reviews (#40)

+++ tests/deleteReview.test.js
@@ -5,12 +5,25 @@ Config.Get(process.env);
 const { AppFactory } = require('../src/app');
 const request = require('supertest');
 
-describe('delete a movie', () => {
+describe('delete a review', () => {
   const mockPgClient = {};
 
-  const app = AppFactory({
-    pgClient: mockPgClient,
-  });
+  const userId = '23';
+
+  const dirtyRottenSessionMiddleware = (req, _, next) => {
+    req.session = {
+      userId: userId,
+    };
+
+    next();
+  };
+
+  const app = AppFactory(
+    {
+      pgClient: mockPgClient,
+    },
+    { middlewares: [dirtyRottenSessionMiddleware] },
+  );
 
   describe('when: a user wants to delete a review', () => {
     let rsp;
@@ -29,8 +42,9 @@ describe('delete a movie', () => {
 
     test('then: we send the correct ID to PgClient AND redirect to reviews page', async () => {
       expect(mockPgClient.dbQuery).toHaveBeenCalledWith(
-        'DELETE FROM reviews WHERE id = $1',
+        'DELETE FROM reviews WHERE id = $1 AND user_id = $2',
         id.toString(),
+        userId,
       );
 
       const status = rsp.status;
